+++ Packages/PimPoPomCore/Sources/PimPoPomCore/GameMode.swift
@@ -1,0 +1,11 @@
+public enum GameMode: String, CaseIterable, Sendable {
+    case arcade
+    case zen
+
+    public var displayName: String {
+        switch self {
+        case .arcade: "Arcade"
+        case .zen: "Zen"
+        }
+    }
+}